Move the user template dir from /pkg/template to /template
