+++ adapter/noop2/BUILD
@@ -11,7 +11,7 @@ go_library(
         "//pkg/adapter:go_default_library",
         "//pkg/adapter/config:go_default_library",
         "//pkg/status:go_default_library",
-        "//pkg/template/sample/report:go_default_library",
+        "//template/sample/report:go_default_library",
         "@com_github_gogo_protobuf//types:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_googleapis_googleapis//:google/rpc",
