+++ pkg/adapter/BUILD
@@ -22,7 +22,7 @@ go_library(
     ],
     deps = [
         "//pkg/adapter/config:go_default_library",
-        "//pkg/template/sample/report:go_default_library",
+        "//template/sample/report:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_googleapis_googleapis//:google/rpc",
         "@com_github_hashicorp_go_multierror//:go_default_library",
