+++ pkg/adapterManager/builderInfoRegistry.gen.go
@@ -19,7 +19,7 @@ import (
 
 	"istio.io/mixer/pkg/adapter"
 	"istio.io/mixer/pkg/adapter/config"
-	sample_report "istio.io/mixer/pkg/template/sample/report"
+	sample_report "istio.io/mixer/template/sample/report"
 )
 
 func DoesBuilderSupportsTemplate(hndlrBuilder config.HandlerBuilder, t adapter.SupportedTemplates, handlerName string) (bool, string) {
