+++ pkg/adapterManager/builderInfoRegistry_test.go
@@ -25,7 +25,7 @@ import (
 
 	"istio.io/mixer/pkg/adapter"
 	"istio.io/mixer/pkg/adapter/config"
-	sample_report "istio.io/mixer/pkg/template/sample/report"
+	sample_report "istio.io/mixer/template/sample/report"
 )
 
 type TestBuilderInfoInventory struct {
