+++ template/sample/report/ReportTesterTemplate.gen.proto
