+++ template/sample/report/ReportTesterTemplate.proto
